requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
Flask-SQLAlchemy==3.0.2
python_dateutil == 2.6.0
psycopg2-binary==2.9.5
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
